refactoring state stats model
