+++ core/templates/services/UpgradedServices.ts
@@ -558,8 +558,6 @@ import { StateSolicitAnswerDetailsService } from
 import { StateSolutionService } from
   // eslint-disable-next-line max-len
   'components/state-editor/state-editor-properties-services/state-solution.service';
-import { StateStatsObjectFactory } from
-  'domain/statistics/StateStatsObjectFactory';
 import { StateTopAnswersStatsBackendApiService } from
   'services/state-top-answers-stats-backend-api.service';
 import { StateTopAnswersStatsObjectFactory } from
@@ -897,7 +895,6 @@ export class UpgradedServices {
       new StateEditorRefreshService();
     upgradedServices['StateGraphLayoutService'] = new StateGraphLayoutService();
     upgradedServices['StateNameService'] = new StateNameService();
-    upgradedServices['StateStatsObjectFactory'] = new StateStatsObjectFactory();
     upgradedServices['StopwatchObjectFactory'] = new StopwatchObjectFactory();
     upgradedServices['StoryNodeObjectFactory'] = new StoryNodeObjectFactory();
     upgradedServices['StoryContentsObjectFactory'] =
@@ -985,8 +982,7 @@ export class UpgradedServices {
       new EndExplorationValidationService(
         upgradedServices['baseInteractionValidationService']);
     upgradedServices['ExplorationStatsObjectFactory'] =
-      new ExplorationStatsObjectFactory(
-        upgradedServices['StateStatsObjectFactory']);
+      new ExplorationStatsObjectFactory();
     upgradedServices['ExplorationTaskObjectFactory'] =
       new ExplorationTaskObjectFactory(
         upgradedServices['HighBounceRateTaskObjectFactory'],
