+++ core/templates/services/exploration-improvements.service.spec.ts
@@ -29,7 +29,7 @@ import { ExplorationStats } from
   'domain/statistics/ExplorationStatsObjectFactory';
 import { PlaythroughObjectFactory } from
   'domain/statistics/PlaythroughObjectFactory';
-import { StateStats } from 'domain/statistics/StateStatsObjectFactory';
+import { StateStats } from 'domain/statistics/state-stats-model';
 import { UserExplorationPermissionsService } from
   'pages/exploration-editor-page/services/user-exploration-permissions.service';
 import { ContextService } from 'services/context.service';
